And/feat#182 challenge detail

+++ android/feature/teacher/challenge/src/main/java/com/sixkids/teacher/challenge/create/ChallengeCreateScreen.kt
@@ -23,7 +23,7 @@ import java.time.LocalDateTime
 @Composable
 fun ChallengeCreateRoute(
     viewModel: ChallengeCreateViewModel = hiltViewModel(),
-    onNavigateResult: (Int, String) -> Unit,
+    onNavigateResult: (Long, String) -> Unit,
     onNavigateUp: () -> Unit,
     onHandleException: (Throwable, () -> Unit) -> Unit,
     onShowSnackbar: (SnackbarToken) -> Unit
